+++ manager/routes/coffeeMarathon.js
@@ -2,7 +2,7 @@ var express = require('express');
 var router = express.Router();
 var async = require('async');
 
-var connection = require('../mysql_connect.js');
+var getSqlConnection = require('../getSqlConnection');
 
 var rowData;
 
@@ -31,69 +31,74 @@ router.get('/', function(req, res, next) {
   async.parallel([
     (callback) => {
       // ユーザランキング作成
-      async.waterfall([
-        (callbackUserRank) => {
-          // ユーザ一覧取得
-          const sql = `select name from ${table_id};`;
-          connection.query(sql, (err, rows) => {
-            if (err) throw err;
-            const userNames = rows.map((data) => {
-              return data.name;
-            });
-            callbackUserRank(null, userNames);
-          })
-        },
-        (userNames, callbackUserRank) => {
-          // ユーザごとの利用料金算出
-          async.map(userNames, (name, callbackMapRowData) => {
-            const sql = `select count(*) from ${table_journal} where name = '${name}' and time >= '${startDate}' and time <= '${endDate}';`;
+      getSqlConnection((err, connection) => {
+        async.waterfall([
+          (callbackUserRank) => {
+            // ユーザ一覧取得
+            const sql = `select name from ${table_id};`;
             connection.query(sql, (err, rows) => {
               if (err) throw err;
-              callbackMapRowData(null, { name, qty: rows[0]['count(*)'] });
+              const userNames = rows.map((data) => {
+                return data.name;
+              });
+              callbackUserRank(null, userNames);
             });
-          }, (err, results) => {
-            if (err) throw err;
-            results.sort((a, b) => {
-              if (a.qty > b.qty) return -1;
-              return 1;
+          },
+          (userNames, callbackUserRank) => {
+            // ユーザごとの利用料金算出
+            async.map(userNames, (name, callbackMapRowData) => {
+              const sql = `select count(*) from ${table_journal} where name = '${name}' and time >= '${startDate}' and time <= '${endDate}';`;
+              connection.query(sql, (err, rows) => {
+                if (err) throw err;
+                callbackMapRowData(null, { name, qty: rows[0]['count(*)'] });
+              });
+            }, (err, results) => {
+              if (err) throw err;
+              results.sort((a, b) => {
+                if (a.qty > b.qty) return -1;
+                return 1;
+              });
+              callbackUserRank(null, results);
             });
-            callbackUserRank(null, results);
-          });
-        },
-      ], (err, results) => {
-        callback(null, results);
+          },
+        ], (err, results) => {
+          connection.release();
+          callback(null, results);
+        });
       });
     },
     (callback) => {
       // 商品別売上げランキングの作成
-      async.waterfall([
-        (callbackDrinkRank) => {
-          // 商品一覧取得
-          const sql = `select name from ${table_drinks}`;
-          connection.query(sql, (err, rows) => {
-            if (err) throw err;
-            const drinkNames = rows.map((data) => {
-              return data.name;
-            });
-            callbackDrinkRank(null, drinkNames);
-          });
-        },
-        (drinkNames, callbackDrinkRank) => {
-          // 商品ごとの売上数算出
-          async.map(drinkNames, (name, callbackMapRowData) => {
-            const sql = `select count(*) from ${table_journal} where item = '${name}' and time >= '${startDate}' and time <= '${endDate}';`;
-            console.log(sql);
+      getSqlConnection((err, connection) => {
+        async.waterfall([
+          (callbackDrinkRank) => {
+            // 商品一覧取得
+            const sql = `select name from ${table_drinks}`;
             connection.query(sql, (err, rows) => {
               if (err) throw err;
-              callbackMapRowData(null, { name, qty: rows[0]['count(*)'] });
+              const drinkNames = rows.map((data) => {
+                return data.name;
+              });
+              callbackDrinkRank(null, drinkNames);
             });
-          }, (err, results) => {
-            callbackDrinkRank(null, results);
-          });
-        },
-      ], (err, results) => {
-        if (err) throw err;
-        callback(null, results);
+          },
+          (drinkNames, callbackDrinkRank) => {
+            // 商品ごとの売上数算出
+            async.map(drinkNames, (name, callbackMapRowData) => {
+              const sql = `select count(*) from ${table_journal} where item = '${name}' and time >= '${startDate}' and time <= '${endDate}';`;
+              connection.query(sql, (err, rows) => {
+                if (err) throw err;
+                callbackMapRowData(null, { name, qty: rows[0]['count(*)'] });
+              });
+            }, (err, results) => {
+              callbackDrinkRank(null, results);
+            });
+          },
+        ], (err, results) => {
+          if (err) throw err;
+          connection.release();
+          callback(null, results);
+        });
       });
     },
   ], (err, results) => {
