+++ manager/views/admin_layout.jade
@@ -13,7 +13,7 @@ html(lang=ja)
     .container
       .logo
         img(src="/images/coffee.png")
-        div.logo-title Vogue cafe
+        div.logo-title Vogue café
 
       .row
         .col-md-3.col-lg-3
