+++ manager/views/coffeeMarathon.jade
@@ -13,7 +13,7 @@ body
   .container
     .logo
       img(src="/images/coffee.png")
-      div.logo-title Vogue cafe
+      div.logo-title Vogue café
 
     .selector
       select.box(name="selectYear")
